fix(nc-gui): show download and view action menu if toolbar operations restricted for that role
