+++ packages/nc-gui/components/smartsheet/toolbar/ViewActionMenu.vue
@@ -8,7 +8,7 @@ const props = withDefaults(
     view: ViewType
     table: TableType
     inSidebar?: boolean
-    showOnlyDownload?: boolean
+    showOnlyCopyId?: boolean
   }>(),
   {
     inSidebar: false,
@@ -240,21 +240,11 @@ defineOptions({
         })
       "
     />
-
-    <template v-if="!view?.is_default && isUIAllowed('viewCreateOrEdit')">
-      <NcDivider />
-      <template v-if="inSidebar">
-        <NcMenuItem v-if="lockType !== LockType.Locked" @click="onRenameMenuClick">
-          <GeneralIcon icon="rename" class="opacity-80" />
-          {{
-            $t('general.renameEntity', {
-              entity: view.type !== ViewTypes.FORM ? $t('objects.view').toLowerCase() : $t('objects.viewType.form').toLowerCase(),
-            })
-          }}
-        </NcMenuItem>
-        <NcTooltip v-else>
-          <template #title> {{ $t('msg.info.disabledAsViewLocked') }} </template>
-          <NcMenuItem disabled>
+    <template v-if="!showOnlyCopyId">
+      <template v-if="!view?.is_default && isUIAllowed('viewCreateOrEdit')">
+        <NcDivider />
+        <template v-if="inSidebar">
+          <NcMenuItem v-if="lockType !== LockType.Locked" @click="onRenameMenuClick">
             <GeneralIcon icon="rename" class="opacity-80" />
             {{
               $t('general.renameEntity', {
@@ -263,27 +253,120 @@ defineOptions({
               })
             }}
           </NcMenuItem>
-        </NcTooltip>
-        <NcMenuItem v-show="lockType !== LockType.Locked" @click="onDescriptionUpdateClick">
-          <GeneralIcon icon="ncAlignLeft" class="opacity-80" />
+          <NcTooltip v-else>
+            <template #title> {{ $t('msg.info.disabledAsViewLocked') }} </template>
+            <NcMenuItem disabled>
+              <GeneralIcon icon="rename" class="opacity-80" />
+              {{
+                $t('general.renameEntity', {
+                  entity:
+                    view.type !== ViewTypes.FORM ? $t('objects.view').toLowerCase() : $t('objects.viewType.form').toLowerCase(),
+                })
+              }}
+            </NcMenuItem>
+          </NcTooltip>
+          <NcMenuItem v-show="lockType !== LockType.Locked" @click="onDescriptionUpdateClick">
+            <GeneralIcon icon="ncAlignLeft" class="opacity-80" />
 
-          {{ $t('labels.editDescription') }}
+            {{ $t('labels.editDescription') }}
+          </NcMenuItem>
+        </template>
+        <NcMenuItem @click="onDuplicate">
+          <GeneralLoader v-if="isOnDuplicateLoading" size="regular" />
+          <GeneralIcon v-else class="nc-view-copy-icon opacity-80" icon="duplicate" />
+          {{
+            $t('general.duplicateEntity', {
+              entity: view.type !== ViewTypes.FORM ? $t('objects.view').toLowerCase() : $t('objects.viewType.form').toLowerCase(),
+            })
+          }}
         </NcMenuItem>
       </template>
-      <NcMenuItem @click="onDuplicate">
-        <GeneralLoader v-if="isOnDuplicateLoading" size="regular" />
-        <GeneralIcon v-else class="nc-view-copy-icon opacity-80" icon="duplicate" />
-        {{
-          $t('general.duplicateEntity', {
-            entity: view.type !== ViewTypes.FORM ? $t('objects.view').toLowerCase() : $t('objects.viewType.form').toLowerCase(),
-          })
-        }}
-      </NcMenuItem>
-    </template>
-    <template v-if="view.type !== ViewTypes.FORM">
-      <NcDivider />
-      <template v-if="isUploadAllowed">
-        <NcSubMenu key="upload" variant="small">
+      <template v-if="view.type !== ViewTypes.FORM">
+        <NcDivider />
+        <template v-if="isUploadAllowed">
+          <NcSubMenu key="upload" variant="small">
+            <template #title>
+              <div
+                v-e="[
+                  'c:navdraw:preview-as',
+                  {
+                    sidebar: props.inSidebar,
+                  },
+                ]"
+                class="nc-base-menu-item group"
+              >
+                <GeneralIcon icon="upload" class="opacity-80" />
+                {{ $t('general.upload') }}
+              </div>
+            </template>
+
+            <NcMenuItemLabel>
+              {{ $t('activity.uploadData') }}
+            </NcMenuItemLabel>
+
+            <template v-for="(dialog, type) in quickImportDialogs">
+              <PermissionsTooltip
+                v-if="isUIAllowed(`${type}TableImport`) && !isPublicView"
+                :key="type"
+                :entity="PermissionEntity.TABLE"
+                :entity-id="table.id"
+                :permission="PermissionKey.TABLE_RECORD_ADD"
+                placement="right"
+                :description="$t('objects.permissions.uploadDataTooltip')"
+              >
+                <template #default="{ isAllowed }">
+                  <NcMenuItem :disabled="!isAllowed" @click="onImportClick(dialog)">
+                    <div
+                      v-e="[
+                        `a:upload:${type}`,
+                        {
+                          sidebar: props.inSidebar,
+                        },
+                      ]"
+                      :class="{ disabled: lockType === LockType.Locked }"
+                      class="nc-base-menu-item"
+                    >
+                      <component
+                        :is="importAlias[type].icon"
+                        v-if="importAlias[type]?.icon"
+                        :class="{ 'opacity-80': isAllowed, '!opacity-50': !isAllowed }"
+                      />
+                      {{ importAlias[type]?.title }}
+                    </div>
+                  </NcMenuItem>
+                </template>
+              </PermissionsTooltip>
+            </template>
+          </NcSubMenu>
+        </template>
+        <NcSubMenu key="download" variant="small">
+          <template #title>
+            <div
+              v-e="[
+                'c:download',
+                {
+                  sidebar: props.inSidebar,
+                },
+              ]"
+              class="nc-base-menu-item group nc-view-context-download-option"
+            >
+              <GeneralIcon icon="download" class="opacity-80" />
+              {{ $t('general.download') }}
+            </div>
+          </template>
+
+          <LazySmartsheetToolbarExportSubActions />
+        </NcSubMenu>
+      </template>
+
+      <template v-if="isUIAllowed('viewCreateOrEdit')">
+        <NcDivider />
+        <NcSubMenu
+          key="lock-type"
+          variant="small"
+          :disabled="!isViewOwner && !isUIAllowed('reAssignViewOwner') && isPersonalView"
+          class="scrollbar-thin-dull max-h-90vh overflow-auto !py-0"
+        >
           <template #title>
             <div
               v-e="[
@@ -292,231 +375,163 @@ defineOptions({
                   sidebar: props.inSidebar,
                 },
               ]"
-              class="nc-base-menu-item group"
+              class="flex flex-row items-center gap-x-3"
             >
-              <GeneralIcon icon="upload" class="opacity-80" />
-              {{ $t('general.upload') }}
+              <div>
+                {{ $t('labels.viewMode') }}
+              </div>
+              <div class="nc-base-menu-item flex !flex-shrink group !py-1 !px-1 rounded-md bg-brand-50">
+                <LazySmartsheetToolbarLockType
+                  :type="lockType"
+                  class="flex nc-view-actions-lock-type !text-brand-500 !flex-shrink !cursor-auto"
+                  hide-tick
+                />
+              </div>
+              <div class="flex flex-grow"></div>
             </div>
           </template>
 
           <NcMenuItemLabel>
-            {{ $t('activity.uploadData') }}
+            {{ $t('labels.viewMode') }}
           </NcMenuItemLabel>
-
-          <template v-for="(dialog, type) in quickImportDialogs">
-            <PermissionsTooltip
-              v-if="isUIAllowed(`${type}TableImport`) && !isPublicView"
-              :key="type"
-              :entity="PermissionEntity.TABLE"
-              :entity-id="table.id"
-              :permission="PermissionKey.TABLE_RECORD_ADD"
-              placement="right"
-              :description="$t('objects.permissions.uploadDataTooltip')"
-            >
-              <template #default="{ isAllowed }">
-                <NcMenuItem :disabled="!isAllowed" @click="onImportClick(dialog)">
+          <NcMenuItem
+            class="!mx-1 !py-2 !rounded-md nc-view-action-lock-subaction max-w-[100px]"
+            data-testid="nc-view-action-lock-subaction-Collaborative"
+            :disabled="!isUIAllowed('fieldAdd')"
+            @click="changeLockType(LockType.Collaborative)"
+          >
+            <SmartsheetToolbarLockType :type="LockType.Collaborative" :disabled="!isUIAllowed('fieldAdd')" />
+          </NcMenuItem>
+          <SmartsheetToolbarNotAllowedTooltip
+            v-if="isEeUI"
+            :enabled="disablePersonalView"
+            :message="isDefaultView ? 'Default view can\'t be made personal' : 'Only view owner can change to personal view'"
+          >
+            <PaymentUpgradeBadgeProvider :feature="PlanFeatureTypes.FEATURE_PERSONAL_VIEWS">
+              <template #default="{ click }">
+                <NcMenuItem
+                  data-testid="nc-view-action-lock-subaction-Personal"
+                  :disabled="disablePersonalView"
+                  class="!mx-1 !py-2 !rounded-md nc-view-action-lock-subaction max-w-[100px] children:w-full children:children:w-full group"
+                  @click="click(PlanFeatureTypes.FEATURE_PERSONAL_VIEWS, () => changeLockType(LockType.Personal))"
+                >
+                  <SmartsheetToolbarLockType
+                    :type="LockType.Personal"
+                    :disabled="disablePersonalView"
+                    @cancel="emits('closeModal')"
+                  />
+                </NcMenuItem>
+              </template>
+            </PaymentUpgradeBadgeProvider>
+          </SmartsheetToolbarNotAllowedTooltip>
+          <NcMenuItem
+            data-testid="nc-view-action-lock-subaction-Locked"
+            class="!mx-1 !py-2 !rounded-md nc-view-action-lock-subaction"
+            :disabled="!isUIAllowed('fieldAdd')"
+            @click="changeLockType(LockType.Locked)"
+          >
+            <SmartsheetToolbarLockType :type="LockType.Locked" :disabled="!isUIAllowed('fieldAdd')" />
+          </NcMenuItem>
+        </NcSubMenu>
+        <template v-if="isEeUI && !isDefaultView">
+          <SmartsheetToolbarNotAllowedTooltip
+            v-if="isPersonalView"
+            :enabled="!(isViewOwner || isUIAllowed('reAssignViewOwner'))"
+            message="Only owner or creator can re-assign"
+          >
+            <PaymentUpgradeBadgeProvider :feature="PlanFeatureTypes.FEATURE_PERSONAL_VIEWS">
+              <template #default="{ click }">
+                <NcMenuItem
+                  inner-class="w-full"
+                  :disabled="!(isViewOwner || isUIAllowed('reAssignViewOwner'))"
+                  @click="click(PlanFeatureTypes.FEATURE_PERSONAL_VIEWS, () => openReAssignDlg())"
+                >
                   <div
                     v-e="[
-                      `a:upload:${type}`,
+                      'c:navdraw:reassign-personal-view',
                       {
                         sidebar: props.inSidebar,
                       },
                     ]"
-                    :class="{ disabled: lockType === LockType.Locked }"
-                    class="nc-base-menu-item"
+                    class="w-full flex flex-row items-center gap-x-3"
                   >
-                    <component
-                      :is="importAlias[type].icon"
-                      v-if="importAlias[type]?.icon"
-                      :class="{ 'opacity-80': isAllowed, '!opacity-50': !isAllowed }"
+                    <div>
+                      {{ $t('labels.reAssignView') }}
+                    </div>
+                    <div class="flex-1 w-full" />
+                    <LazyPaymentUpgradeBadge
+                      :feature="PlanFeatureTypes.FEATURE_PERSONAL_VIEWS"
+                      :limit-or-feature="'to access re-assign personal view feature.' as PlanFeatureTypes"
+                      :content="
+                        $t('upgrade.upgradeToAccessReassignViewSubtitle', {
+                          plan: getPlanTitle(PlanTitles.PLUS),
+                        })
+                      "
+                      :on-click-callback="() => emits('closeModal')"
                     />
-                    {{ importAlias[type]?.title }}
                   </div>
                 </NcMenuItem>
               </template>
-            </PermissionsTooltip>
-          </template>
-        </NcSubMenu>
-      </template>
-      <NcSubMenu key="download" variant="small">
-        <template #title>
-          <div
-            v-e="[
-              'c:download',
-              {
-                sidebar: props.inSidebar,
-              },
-            ]"
-            class="nc-base-menu-item group nc-view-context-download-option"
+            </PaymentUpgradeBadgeProvider>
+          </SmartsheetToolbarNotAllowedTooltip>
+          <SmartsheetToolbarNotAllowedTooltip
+            v-else
+            :enabled="!isViewOwner"
+            message="Only view owner can assign as personal view"
           >
-            <GeneralIcon icon="download" class="opacity-80" />
-            {{ $t('general.download') }}
-          </div>
-        </template>
-
-        <LazySmartsheetToolbarExportSubActions />
-      </NcSubMenu>
-    </template>
-
-    <template v-if="isUIAllowed('viewCreateOrEdit')">
-      <NcDivider />
-      <NcSubMenu
-        key="lock-type"
-        variant="small"
-        :disabled="!isViewOwner && !isUIAllowed('reAssignViewOwner') && isPersonalView"
-        class="scrollbar-thin-dull max-h-90vh overflow-auto !py-0"
-      >
-        <template #title>
-          <div
-            v-e="[
-              'c:navdraw:preview-as',
-              {
-                sidebar: props.inSidebar,
-              },
-            ]"
-            class="flex flex-row items-center gap-x-3"
-          >
-            <div>
-              {{ $t('labels.viewMode') }}
-            </div>
-            <div class="nc-base-menu-item flex !flex-shrink group !py-1 !px-1 rounded-md bg-brand-50">
-              <LazySmartsheetToolbarLockType
-                :type="lockType"
-                class="flex nc-view-actions-lock-type !text-brand-500 !flex-shrink !cursor-auto"
-                hide-tick
-              />
-            </div>
-            <div class="flex flex-grow"></div>
-          </div>
-        </template>
-
-        <NcMenuItemLabel>
-          {{ $t('labels.viewMode') }}
-        </NcMenuItemLabel>
-        <NcMenuItem
-          class="!mx-1 !py-2 !rounded-md nc-view-action-lock-subaction max-w-[100px]"
-          data-testid="nc-view-action-lock-subaction-Collaborative"
-          :disabled="!isUIAllowed('fieldAdd')"
-          @click="changeLockType(LockType.Collaborative)"
-        >
-          <SmartsheetToolbarLockType :type="LockType.Collaborative" :disabled="!isUIAllowed('fieldAdd')" />
-        </NcMenuItem>
-        <SmartsheetToolbarNotAllowedTooltip
-          v-if="isEeUI"
-          :enabled="disablePersonalView"
-          :message="isDefaultView ? 'Default view can\'t be made personal' : 'Only view owner can change to personal view'"
-        >
-          <PaymentUpgradeBadgeProvider :feature="PlanFeatureTypes.FEATURE_PERSONAL_VIEWS">
-            <template #default="{ click }">
-              <NcMenuItem
-                data-testid="nc-view-action-lock-subaction-Personal"
-                :disabled="disablePersonalView"
-                class="!mx-1 !py-2 !rounded-md nc-view-action-lock-subaction max-w-[100px] children:w-full children:children:w-full group"
-                @click="click(PlanFeatureTypes.FEATURE_PERSONAL_VIEWS, () => changeLockType(LockType.Personal))"
-              >
-                <SmartsheetToolbarLockType
-                  :type="LockType.Personal"
-                  :disabled="disablePersonalView"
-                  @cancel="emits('closeModal')"
-                />
-              </NcMenuItem>
-            </template>
-          </PaymentUpgradeBadgeProvider>
-        </SmartsheetToolbarNotAllowedTooltip>
-        <NcMenuItem
-          data-testid="nc-view-action-lock-subaction-Locked"
-          class="!mx-1 !py-2 !rounded-md nc-view-action-lock-subaction"
-          :disabled="!isUIAllowed('fieldAdd')"
-          @click="changeLockType(LockType.Locked)"
-        >
-          <SmartsheetToolbarLockType :type="LockType.Locked" :disabled="!isUIAllowed('fieldAdd')" />
-        </NcMenuItem>
-      </NcSubMenu>
-      <template v-if="isEeUI && !isDefaultView">
-        <SmartsheetToolbarNotAllowedTooltip
-          v-if="isPersonalView"
-          :enabled="!(isViewOwner || isUIAllowed('reAssignViewOwner'))"
-          message="Only owner or creator can re-assign"
-        >
-          <PaymentUpgradeBadgeProvider :feature="PlanFeatureTypes.FEATURE_PERSONAL_VIEWS">
-            <template #default="{ click }">
-              <NcMenuItem
-                inner-class="w-full"
-                :disabled="!(isViewOwner || isUIAllowed('reAssignViewOwner'))"
-                @click="click(PlanFeatureTypes.FEATURE_PERSONAL_VIEWS, () => openReAssignDlg())"
-              >
-                <div
-                  v-e="[
-                    'c:navdraw:reassign-personal-view',
-                    {
-                      sidebar: props.inSidebar,
-                    },
-                  ]"
-                  class="w-full flex flex-row items-center gap-x-3"
+            <PaymentUpgradeBadgeProvider :feature="PlanFeatureTypes.FEATURE_PERSONAL_VIEWS">
+              <template #default="{ click }">
+                <NcMenuItem
+                  inner-class="w-full"
+                  :disabled="!isViewOwner"
+                  @click="click(PlanFeatureTypes.FEATURE_PERSONAL_VIEWS, () => openReAssignDlg())"
                 >
-                  <div>
-                    {{ $t('labels.reAssignView') }}
+                  <div
+                    v-e="[
+                      'c:navdraw:assign-personal-view',
+                      {
+                        sidebar: props.inSidebar,
+                      },
+                    ]"
+                    class="w-full flex flex-row items-center gap-x-3"
+                  >
+                    <div>
+                      {{ $t('labels.assignAsPersonalView') }}
+                    </div>
+                    <div class="flex-1 w-full" />
+                    <LazyPaymentUpgradeBadge
+                      :feature="PlanFeatureTypes.FEATURE_PERSONAL_VIEWS"
+                      :limit-or-feature="'to access assign as personal view feature.' as PlanFeatureTypes"
+                      :content="
+                        $t('upgrade.upgradeToAccessAssignAsPersonalViewSubtitle', {
+                          plan: getPlanTitle(PlanTitles.PLUS),
+                        })
+                      "
+                      :on-click-callback="() => emits('closeModal')"
+                    />
                   </div>
-                  <div class="flex-1 w-full" />
-                  <LazyPaymentUpgradeBadge
-                    :feature="PlanFeatureTypes.FEATURE_PERSONAL_VIEWS"
-                    :limit-or-feature="'to access re-assign personal view feature.' as PlanFeatureTypes"
-                    :content="
-                      $t('upgrade.upgradeToAccessReassignViewSubtitle', {
-                        plan: getPlanTitle(PlanTitles.PLUS),
-                      })
-                    "
-                    :on-click-callback="() => emits('closeModal')"
-                  />
-                </div>
-              </NcMenuItem>
-            </template>
-          </PaymentUpgradeBadgeProvider>
-        </SmartsheetToolbarNotAllowedTooltip>
-        <SmartsheetToolbarNotAllowedTooltip v-else :enabled="!isViewOwner" message="Only view owner can assign as personal view">
-          <PaymentUpgradeBadgeProvider :feature="PlanFeatureTypes.FEATURE_PERSONAL_VIEWS">
-            <template #default="{ click }">
-              <NcMenuItem
-                inner-class="w-full"
-                :disabled="!isViewOwner"
-                @click="click(PlanFeatureTypes.FEATURE_PERSONAL_VIEWS, () => openReAssignDlg())"
-              >
-                <div
-                  v-e="[
-                    'c:navdraw:assign-personal-view',
-                    {
-                      sidebar: props.inSidebar,
-                    },
-                  ]"
-                  class="w-full flex flex-row items-center gap-x-3"
-                >
-                  <div>
-                    {{ $t('labels.assignAsPersonalView') }}
-                  </div>
-                  <div class="flex-1 w-full" />
-                  <LazyPaymentUpgradeBadge
-                    :feature="PlanFeatureTypes.FEATURE_PERSONAL_VIEWS"
-                    :limit-or-feature="'to access assign as personal view feature.' as PlanFeatureTypes"
-                    :content="
-                      $t('upgrade.upgradeToAccessAssignAsPersonalViewSubtitle', {
-                        plan: getPlanTitle(PlanTitles.PLUS),
-                      })
-                    "
-                    :on-click-callback="() => emits('closeModal')"
-                  />
-                </div>
-              </NcMenuItem>
-            </template>
-          </PaymentUpgradeBadgeProvider>
-        </SmartsheetToolbarNotAllowedTooltip>
+                </NcMenuItem>
+              </template>
+            </PaymentUpgradeBadgeProvider>
+          </SmartsheetToolbarNotAllowedTooltip>
+        </template>
       </template>
-    </template>
 
-    <template v-if="!view.is_default && isUIAllowed('viewCreateOrEdit')">
-      <NcDivider />
-      <NcTooltip v-if="lockType === LockType.Locked">
-        <template #title> {{ $t('msg.info.disabledAsViewLocked') }} </template>
-        <NcMenuItem disabled>
+      <template v-if="!view.is_default && isUIAllowed('viewCreateOrEdit')">
+        <NcDivider />
+        <NcTooltip v-if="lockType === LockType.Locked">
+          <template #title> {{ $t('msg.info.disabledAsViewLocked') }} </template>
+          <NcMenuItem disabled>
+            <GeneralIcon class="nc-view-delete-icon opacity-80" icon="delete" />
+            {{
+              $t('general.deleteEntity', {
+                entity:
+                  view.type !== ViewTypes.FORM ? $t('objects.view').toLowerCase() : $t('objects.viewType.form').toLowerCase(),
+              })
+            }}
+          </NcMenuItem>
+        </NcTooltip>
+        <NcMenuItem v-else danger @click="onDelete">
           <GeneralIcon class="nc-view-delete-icon opacity-80" icon="delete" />
           {{
             $t('general.deleteEntity', {
@@ -524,20 +539,12 @@ defineOptions({
             })
           }}
         </NcMenuItem>
-      </NcTooltip>
-      <NcMenuItem v-else danger @click="onDelete">
-        <GeneralIcon class="nc-view-delete-icon opacity-80" icon="delete" />
-        {{
-          $t('general.deleteEntity', {
-            entity: view.type !== ViewTypes.FORM ? $t('objects.view').toLowerCase() : $t('objects.viewType.form').toLowerCase(),
-          })
-        }}
-      </NcMenuItem>
+      </template>
     </template>
   </NcMenu>
   <span v-else v-bind="$attrs"></span>
 
-  <template v-if="table?.base_id && currentSourceId">
+  <template v-if="table?.base_id && currentSourceId && isUploadAllowed">
     <!-- Don't add this inside the NcMenu else it will show 2 modals at the same time -->
     <LazyDlgQuickImport
       v-for="tp in quickImportDialogTypes"
